requirements.txt
allure-pytest==2.8.16
allure-python-commons==2.8.16
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.3.1
html-testRunner==1.2.1
idna==2.9
importlib-metadata==1.6.1
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.8.2
pycparser==2.20
pyparsing==2.4.7
pytest==5.4.3
requests==2.24.0
selenium==3.141.0
six==1.15.0
unittest-data-provider==1.0.1
urllib3==1.25.9
wcwidth==0.2.4
webdriver-manager==3.2.1
zipp==3.1.0
